# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JJK Mission Mailer API",
        "description": "API for managing sorcerer missions and assignments",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/curse/{id}": {
            "get": {
                "tags": [
                    "Curses"
                ],
                "summary": "Get a curse by ID",
                "description": "Returns a single curse with related missions",
                "operationId": "d045883f7ed02d3585a63625dab42038",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Curse details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Curse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Curse not found"
                    }
                }
            }
        },
...
        "/assignment/{id}": {
            "get": {
                "tags": [
                    "Assignments"
                ],
                "summary": "Get an assignment by ID",
                "description": "Returns a single mission assignment with full details",
                "operationId": "85f8e70e2afc203e7908754602324f76",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assignment details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MissionAssignment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Assignment not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Curse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Finger Bearer"
                    },
                    "description": {
                        "type": "string"
                    },
                    "curse_level": {
                        "type": "string",
                        "example": "special-grade"
                    },
                    "curse_type": {
                        "type": "string",
                        "example": "Cursed Spirit"
                    },
                    "abilities": {
                        "type": "string"
                    },
                    "known_weaknesses": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "example": "at_large"
                    },
                    "first_sighted_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "MissionAssignment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "mission_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sorcerer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "assigned_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "result_status": {
                        "type": "string",
                        "enum": [
                            "success",
                            "partial_success",
                            "failure",
                            "aborted"
                        ],
                        "nullable": true
                    },
                    "casualties": {
                        "type": "integer",
                        "example": 0
                    },
                    "mission_report": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}